// src/generated/epr-flow-control-api/model/reportResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * ERP Flow Control API
 * The ERP Flow Control API is responsible for managing the lifecycle of reports.

 * OpenAPI spec version: 1.0
 */
import type { ReportResponseDtoReportStatus } from './reportResponseDtoReportStatus';

/**
 * Report Response
 */
export interface ReportResponseDto {
  /** Correlation ID */
  correlationId?: string;
  /** Code UC */
  codeUc?: string;
  /** Nickname */
  nickname?: string;
  /** Reference date */
  referenceDate?: string;
  /** Economic group */
  economicGroup?: string;
  /** Report Status */
  reportStatus?: ReportResponseDtoReportStatus;
  /** Message */
  message?: string;
  /** Updated at */
  updatedAt?: string;
  /** Ingested at */
  ingestedAt?: string;
  /** Built at */
  builtAt?: string;
  /** Sent at */
  sentAt?: string;
}