// src/generated/epr-flow-control-api/model/searchResultResponseReportResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * ERP Flow Control API
 * The ERP Flow Control API is responsible for managing the lifecycle of reports.

 * OpenAPI spec version: 1.0
 */
import type { PageResponse } from './pageResponse';
import type { ReportResponseDto } from './reportResponseDto';

export interface SearchResultResponseReportResponseDto {
  response?: ReportResponseDto[];
  paging?: PageResponse;
}